+++ front/app/api/graph_data_units/responseTypes/index.ts
@@ -1,8 +1,4 @@
 import {
-  CommentsCountRow,
-  TimeSeriesResponseRow as CommentsTimeSeriesResponseRow,
-} from 'components/admin/GraphCards/CommentsByTimeCard/useCommentsByTime/typings';
-import {
   InputsCountRow,
   TimeSeriesResponseRow as PostsTimeSeriesResponseRow,
 } from 'components/admin/GraphCards/PostsByTimeCard/usePostsByTime/typings';
@@ -15,13 +11,6 @@ export type PostsByTimeResponse = {
   data: {
     type: 'report_builder_data_units';
     attributes: [PostsTimeSeriesResponseRow[] | [], [InputsCountRow] | []];
-  };
-};
-
-export type CommentsByTimeResponse = {
-  data: {
-    type: 'report_builder_data_units';
-    attributes: [CommentsTimeSeriesResponseRow[] | [], [CommentsCountRow] | []];
   };
 };
